Add --output-package-location flag to dump JSON of location
Summary:
Our PyPI publishing script no longer works due to (probably) a new bug in Eden. This fixes that issue by using the build in /tmp rather than moving the result into Eden-backed directories.

This diff dumps the dataclass when the `--output-artifact-location` flag is passed in. We don't use this internally (as far as I can tell), so it doesn't matter much how we return this information.

Reviewed By: grievejia

Differential Revision: D61341113

fbshipit-source-id: 159ccab3e6f7e7d17d8efd5f3b45bfef3cee04dc