Add a comment clarifying invalid type variables for tuple
Summary:
I was looking at this code trying to figure out how Pyre knows how
to zip a tuple's parameters and variables for a while before I realized
that this is the wrong question - Pyre uses a special form for
generic tuples, so only the Primitive case matters here.

I figured since this is a very low-level code detail that couples lines
5421-5432 of `type.ml` pretty tightly to lines 2412-2414 of
`attributeResolution.ml` (as in, if you don't know *exactly* what the former
is doing then the latter appears to be suspect), it maybe deserved a
clarifying comment on how the two chunks of logic fit together.

Reviewed By: yangdanny97

Differential Revision: D61250884

fbshipit-source-id: 463f01c55e2653ca254d576ecbf2b98b081329de