Strip `resolve_literal` from all public APIs
Summary:
The `resolve_literal` function is a helper in attributeResolution.ml that
is used in two ways:
- to handle some cases of resolving a global (in particular, it handles
  globals with no annotation, where we can *sometimes* infer the type)
- to handle some cases of figuring out an attribute in an attribute
  table; I'm actually not 100% clear when Pyre allows implicit types
  with no annotation (I think we're stricter than with globals) but it does get
  called in some cases.

But it is *never* used directly - it's exposed exclusively through attribute table
and global signature analysis. We should get rid of it from the public API.

This change was motivated partly because it does need a `~variable_map` for scoped
type vars (because in the attribute table case there might be type vars in scope
from containing classes / functions), but in the global case it would never need
scoped tyep vars. I was going to try to massage the API to somehow reflect this,
and then I realized that it's actually unneeded.

Reviewed By: yangdanny97

Differential Revision: D61337081

fbshipit-source-id: af47da848b2faa5bab44014846cbb21d05cf6235