Rename Type.Parameter to Type.Argument, part 1
Summary:
**This stack of 2 diffs**

The `Type.t` data in OCaml primarily represents a (possibly propagated through
analysis) type *annotation*. It does not represent in any meaningful sense a
type *constructor*, which is a concept that has no direct representation in
Pyre but is basically mapped to a "class" in the current architecture.

In terms of modern vocabulary, typically we talk about "parameters" versus
"arguments"; apparently there's also a mostly older vocabulary that calls these
"formal parameters" vs "parameters", but "paramters" / "arguments" is already
the vocabulary that Pyre uses for function calls and type instantiations have
the same distinction.

But because a `Type.t` loosely corresponds to an *annotation*, the values
inside of a `Type.Parametric` form are not actually "parameters", they are
"arguments".

I have two motivations for making this change now:
- We're starting to have discussions about the architecture of how Pyre
  represents types, and it's confusing to distinguish argument vs parameter
  in discussions (which is important, if you cannot name a concept then you
  can't really have a conversation about it) and yet have the code use
  the wrong name. If the fix is easy, why not fix the code?
- In order to improve our handling of type constructors (for now, generic
  classes) when supporting PEP 695, we will probably need a cleaner notion of
  type *parameters*, which means I want that name to be available so that
  we have the option of using it in a refactor.
  - for existing generics, the type *parameters* will be extracted from the
    type *arguments* of the "generic base class"
  - but for PEP 695 classes this will not be the case, the type parameters
    will be explicitly marked as they are in most other languages. This is part
    of why confusing the two concepts is such a problem for 695.

**This diff**

Does the easy part, renaming just the module.

Renaming variables and functions that are confusing `argument` with `parameter` is tricker and in a separate stacked diff.

Reviewed By: migeed-z

Differential Revision: D61548862

fbshipit-source-id: b7724e8b0b511ef3270d26723c8390515bfd3a53