Explain implicit generic alias handling
Summary:
The exact behavior of `replace_unbound_parameters_with_any`
took me several hours to figure out here - it's fairly obvious
*what* it does, but rather difficult to figure out *why* and
what the actual implications are for Pyre as a system.

I'm fairly certain I did finally work it out - this flag is here
to handle implicitly generic aliases. The way it gets handled
is probably not really intended, because a structurally illegal
type alias gets treated as equivalent to an implicitly generic
alias, although it's unclear how much this matters (if the
structurally invalid alias is in a type checked module, we would
still throw a type error on its definition).

Reviewed By: grievejia

Differential Revision: D61431061

fbshipit-source-id: 3adf0983dfd727592c20809332c7d024653d1776