De-obfuscate class name fetch in ClassHierarchy
Summary:
I was reading this code to understand the extent to which it
really depends on having base classes available as `Type.t`
values (versus whether some less-fully-resolved data would
suffice here).

This line of code was misleading - it looked like the MRO
resolution maybe depended on creating a proper `Type.t` for
a parent, but this was just bad naming, all it does is get
the class name which explicitly does *not* depend on anything
other than the name of the type constructor (i.e. the outermost
name of a parametric annotation).

Renaming to clarify this.

Reviewed By: migeed-z

Differential Revision: D61485470

fbshipit-source-id: 5fab0b1b11d490cc6eb705039781a50ea589e2c4