modularize locationBasedLookup.ml
Summary:
For quite some time now I've had a very hard time skimming the logic
in locationBasedLookup because several different things are mixed
together in one module:
- computing expression types using a location table
- computing expression level coverage, which is a significant chunk
  of logic downstream of ^^
- doing fine-grained symbol selection, which is a dijoint code path
  from the full-file lookups above
- implementing IDE related features and other end-user queries on top
  of the symbol selection logic

The intermingling of concerns, especially given that fine-grained
symbol selection logic is share-nothing with full-file coverage (other
than a few of the datatypes are shared with expression level coverage),
was making it very hard for me to see what was going on, and even to
know exactly when we use the fine-grained symbol finding vs the
type x location lookup table.

This commit makes no actual code changes, but modularizes everything:
- ExpressionTypes has the basic logic for all types at locations per file
- LocationWithSymbol has some shared datatypes used in both expression
  level coverage and single-symbol queries. It has almost no actual logic.
- ExpressionLevelCoverage has the (downstream of ^^) logic to tag
  coverage gaps; it is used both by the IDE integration and by `pyre report`
- SymbolSelection has the core code for finding a symbol given a position
  or location, which is at the heart of most of the lightweight IDE queries
- SingleSymbolQueris has the end-user queries (most of which correspond to
  some kind of LSP functionality) downstream of SymbolSelection

This reorganization is a huge BE win for me, because for ongoing work with a
static analysis partner team I care almost exclusively about the very topmost
`ExpressionTypes` module, and this diff makes it much easier to isolate
and improve just that code without being confused by all of the downstream
use cases.

Reviewed By: grievejia

Differential Revision: D61608504

fbshipit-source-id: 02d6bbacd4c9ca88f899bc20b4b6b9e09dd132df