Fix record field name
Summary:
On an intermediate diff in the stack introducing this concept,
we were storing the "generic base type arguments" inside
of `Edges.t`

But a later diff cut this over to storing the `Type.Variable.t`
values themselves, and we didn't change the record field name.

This is a quick fix to avoid confusion; we don't want something
that isn't `Type.Argument.t` based named "arguments".

Reviewed By: grievejia

Differential Revision: D61666667

fbshipit-source-id: 83eaa21dee22f1be9f6cc8583205d8da3b0b135e