Use tuple type information for more precise `*args` unpacking
Summary:
When `*args` is passed to a function, if the type of `args` is a fixed-length tuple type, then we can expand `*args` into a fixed number of arguments even if the value of `args` is not known to be constant.

When doing this expansion, the positions of all subsequent arguments need to be updated. This fixes some incorrect "nth positional argument" counting in error messages.

Reviewed By: stroxler

Differential Revision: D61581142

fbshipit-source-id: 9529615a3414de91785121ccb1e4973554c1861f