Partition source and sink trees by kind before checking triggered flows
Summary:
Currently, the time complexity of function `compute_triggered_flows` is high --
it is the product of:
1. number of tuples of source kinds and source call info (from the source tree)
2. number of tuples of sink kinds, sink call info, and frames (from the sink
tree)
3. number of triggers, when the partial sink in (2) is triggered
4. the time complexity of calling `generate_issue_or_attach_subtraces`

In addition, function `compute_triggered_flows` is called for `n` times, if a
call site has `n` arguments.

To speed this up, we could
1. filter away non-partial sinks, so that (2) above becomes smaller.
2. avoid explicitly extracting out tuples of sources and call_infos (or sinks
call_infos, and frames). Instead, since only the taint kinds are "important"
(because they are used to generate flows), we could partition the taint trees
by taint kinds, so that the sizes of (1) and (2) both become smaller.

Implementation-wise, this diff also
1. updates the type of `Trigger` and the value type of map
`TriggeredSinkForCall` into a map from `Trigger`s to the subtraces (as opposed
to a set of `Trigger`s), so that we can further deduplicate `Triggers` due to
only distinguishing them based on the triggering source and the argument
location.
2. breaks down function `generate_issue_or_attach_subtraces` into
`generate_issue` and `attach_subtraces`, to improve the readability.

Reviewed By: arthaud

Differential Revision: D61672566

fbshipit-source-id: aed68ac360d31e9b8cb086647179138198c7e1f5