Eliminate the use of `parents` in determining type parameters
Summary:
Up until this diff, the determination of type parameters lived in `ClassHierarchy`
but it didn't really have a clear representation in data:
- it implicitly relied on the assumption that all type parameters are expressed as
  type arguments on a generic base class, which is inconsistent with what we
  need for PEP 695 support,
- it also relied on a lot of indirection by way of `parents_and_generic_of_target`,
   which was folding the *actual* parents together with a synthesized generic
   base class and then performing an algorithm that could depend on both

This diff does the work needed to actually directly represent the parameters as
data that we can extract in `ClassHierarchyEnvironment.produce_value`. It turns out
that because of Pyre's reliance in special forms as "fake classes" inheriting from
`Generic[Any]` we have to explicitly distinguish three different cases:
- a non-generic class (no generic base)
- a legacy-style generic class with a generic base that has type vars
- a class with an invalid generic base (someone tried to pass `Generic[C]` for a
   type `C` that wasn't a type variable... e.g. we ourselves hardcode `Generic[Any]`)

This is messy, but also not something we need to urgently fix to keep PEP 695 support
moving so I filed T199653412 to deal with it later.

**What's next?**

Note that while we now have a closed-form representation of type parameters, we certainly
are not finished yet: the parameters are being represented as `Type.Variable.t`, and we
are relying on the fact that these are fully-formed types with well-defined variance. This
is not consistent with PEP 695. So there is more work to do, this diff is mostly just making
it clear where we need to make more datatype changes as we go.

We probably want to use something more like `Type.Variable.Declaration.t`, and eventually
we'll probably need the variance to be possibly `NotYetInferred`

Reviewed By: migeed-z

Differential Revision: D61685056

fbshipit-source-id: 96be493718520a19d6d00a6fb468bfa9ef4f2120