Qualify type statements
Summary:
In an effort towards supporting PEP695 syntax, we are supporting type statements, which involves collecting them in the UGE, as shown in the previous diff, and then collecting them in the typeAliasEnvironment layer.

Due to qualification, we must qualify type statements before collecting them in UGE. This diff does this. In the next diff, we add unittests for this.

Reviewed By: grievejia

Differential Revision: D61742334

fbshipit-source-id: b8328c050fe6d8d4e4340a19b8b3e44a13af1c32