Make ParseAnnotationCache PEP 695-friendly [2] toggle the cache
Summary:
In the previous diff, I made `scoped_type_variables` plain data that could
be passed, so that by passing `None` we could signal to `parse_annotation`
that no scoped type variables matter on the current annotation.

The motivation was to allow us to proceed with PEP 695 syntax support, which
isn't friendly to `ParseAnnotationCache`, without affecting production
performance: we can skip the cache but only when there are scoped type variables,
which should have no impact on prod performance until codebases actually begin
to use the new syntax.

The goal here is not to let us avoid solving the perf issues but rather to
allow development of the *features* to proceed "as if there were no cache" while
we work on a strategy to avoid the perf hit and actually remove the cache.

------

With servicelab broken it's a little hard to see perf, but I've rebased 4 times to get 4
Pysa signals, and the typeCheck phase has been about a minute faster twice and about
a minute slower twice. This suggests that the impact of the extra object creation here
on perf is pretty small (and also indicates that a +/- 15% variation in type check turnaround
of one pysa distillery run is pretty common, which is good to know for future reference!)

Reviewed By: migeed-z

Differential Revision: D61702011

fbshipit-source-id: a5b62bad0dd29d4c86ec7112beb2c8ce5792c748