Make ParseAnnotationCache PEP 695-friendly [1] use scoped_type_variables as data
Summary:
In order to work around the fact that PEP 695 *feature* development is currently
blocked on the problem that the ParseAnnotationCache isn't very scoped-type-variable
friendly, we want a temporary work around where we make `parse_annotation` actually
know whether it's using scoped type variables or only legacy-style typevars.

This will allow us to skip the cache only when there are scoped type variables, which
will in turn unblock feature development because thee won't be a perf cost on
existing codebases that don't use the new syntax.

We want to do this so that the feature work can proceed at the same time as we
investigate how to mitigate the performance impact of losing this cache. It's a temporary
fix because once a codebase relies heavily on scoped type varaibles the cache becomes
useless and the perf problem (slowly) appears in production, but it should really
improve our ability to move quickly on this, given that perf investigations tend to
take time.

This diff is the first of two that will allow us to accomplish this: it cuts all of
the `attributeResolution.ml` logic over from taking an abstract `variable_map` that
combines both scoped and legacy type vars to instead take `scoped_type_variables`
directly as an optional map.

The next diff will then skip the cache when that map is nonempty (which it never
is right now), at which point feature development can prodceed with no impact
on production perf against codebases that don't use PEP 695 syntax.

NOTE: as of this diff I am using either scoped type variables or legacy type varaibles
but never both in `parse_annotation`. It is not clear whether this will acutally work,
we may need to use fallback logic that lets users mix the two (it's a type error to
mix them, but that may not mean it's a good idea for Pyre to not be best-effort in
that case). It's easy to pivot on this later but good to be aware that I made this
choice for now cc migeed-z

Reviewed By: migeed-z

Differential Revision: D61702012

fbshipit-source-id: de979b96356e668f0b6c95c0941c2c434d6794a5