Strip `?replace_unbound_parameters_with_any` from typeAliasEnvironment
Summary:
This flag is meaningless upstream of `attributeResolution.ml` logic because the
only code that interprets the flag (the only code even *capable* of it) lives
there, and depends heavily on `classHierarchyEnvironment` for "what are the type
parameters"?

Stripping this flag from the TAE query functions is a strict win; it simplifies
the code by making interfaces easier to read and eliminating a bunch of
`replace_unbound_parameters_with_any:_` junk parameters, and it also makes it
so that said interfaces are no longer actively misleading about what that
flag actually does and where in the stack that operation is able to happen.

Reviewed By: yangdanny97

Differential Revision: D61772330

fbshipit-source-id: b5927801c58a4715e3def6fd9fe09f51f791730c