Rename VariableAlias -> VariableDeclaration
Summary:
The typeAliasEnvironment data for a variable declaration stores
a `Type.Variable.Declaration.t`; I actually think I may want to rename
the latter type so that it can be used for PEP 695 type vars as well,
but the typeAliasEnvironment data is always representing a decalartion
(where by "declaration" I mean a *statement* that is declaring a type
variable *name* to be referrable from the current scope (thereby creating an
implicitly scoped type var on the using function or class); in this sense I
intend the name to evoke the notion of a C language "declaration").

Given this, a rename from `VariableAlias` to `VariableDeclaration` makes
the name more clearly map to the underlying concept and data.

Reviewed By: yangdanny97

Differential Revision: D61806841

fbshipit-source-id: 3ceaafbdb5219c52d73fc33abf39b726239dcc67