Kill unused `empty_alliases` function, clean up typeTest some
Summary:
In the current codebase, Pyre expects to get type aliases and variables
separately. The tests are in a weird state where a lot of the logic is
based on older patterns, and it's forcing us to make `RawAlias.t`
public., which at the moment I'd prefer not to do (I'm very uncertain
about the long-term approach - type aliases and other notions of
aliases/rebinds are actually pretty complex in Python).

This commit cleans up `typeTest`, which was often jumping through hoops
to make a map from name to `Type.t`, then wrapping it in a `TypeAlias`,
then unwrapping it again which is kind of silly.

The payoff in production interfaces is that we can remove the `empty_aliases`
function, which has no legitimate use anymore.

Reviewed By: yangdanny97

Differential Revision: D61806842

fbshipit-source-id: 2e6222e1eb2e735d29a2ee749137260e76dd241e