Add a test case for nesting behavior in FDE
Summary:
This diff kicks off an effort to support access for a downstream
static analysis team that wants to be able to query
`callees_with_location(...)` on nested functions.

To do this I plan to use the new `ModuleContext.t` data Jia has defined
as part of his qualification efforts to get a non-mangled qualified name
for each define.

This change won't provide access to nested *classes*, which are currently
dropped from type checking (we could probably reverse this) but it should
provide access to nested *functions*.

This commit just adds a test case illustrating how FDE currently models the
names of functions, which will make it easy for me to validate the
behavior change when I cut over.

Reviewed By: grievejia

Differential Revision: D61748414

fbshipit-source-id: 11f569c1ae6ebd73f15cdf09bf9a3d36922f0fda