Update PyPI script to accept output directory to move artifacts to
Summary:
My changes fixed the issue with linux CI packaging, but now OSX is showing the same issue that was happening, where the file can't be found but should be accessible, likely due to an Eden issue.

This diff updates the OSS PyPI build package logic to now take in an output directory to move the artifacts to.

Reviewed By: grievejia

Differential Revision: D61802335

fbshipit-source-id: 03a4068e52e6deb77faf1ace70184a488d3f27a2