Add CustomAvailability availability type
Summary:
I'd like to add new functionality around reloading GK values occasionally while Pyre is running in the IDE. This will allow us to quickly switch features on and off or change the rollout without having to wait until users restart VSCode.

This diff adds an Availability type that uses custom logic to determine if a feature is enabled or disabled.

Reviewed By: grievejia

Differential Revision: D61686856

fbshipit-source-id: 640be14e72ded4582b6366e60eafd53065d73e80