Delete unused `attribute_names`
Summary:
The attributeResolution.ml code is one of our biggest headaches
maintaining Pyre because of the sheer number of mutually recursive
methods.

In addition, if we are trying to understand perf it becomes valuable
to trace all calls that can trigger resolving the full transitive attribute
table of a class because that is pretty expensive.

This one is unused, let's kill it!

Reviewed By: connernilsen

Differential Revision: D61729229

fbshipit-source-id: c7e8d54c4e905b9f0a2a648e8f4738bb5f282c4d