Move and simplify `attribute_details` logic
Summary:
The `AttributeDetail` datatype and the `attribute_details` query in
`attributeResolution.ml` are not actually part of Pyre's analysis engine at all,
they are used *exclusively* in the IDE-oriented logic that lives in
`LocationBasedLookup.SingleSymbolQueries`, and can be expressed cleanly in
terms of already well-abstracted attribute table operations.

So lets move the logic to where it is used; this reduces the surface area of
globalResolution which is great, but much more important is that it eliminates
another of the recursive methods in `AttributeResolution`'s `base` class.

In addition there were flags on this query that have no existing use case. Since
the IDE support is currently paused anyway, I see no reason to thread this unused
complexity through at this time; I simplified a bit by removing unused arguments.

Reviewed By: yangdanny97

Differential Revision: D61741836

fbshipit-source-id: 03b408bc213a4b0d3b8b305f7d529a3764207c58