Add new isolation dir for PTT to test out type check durations
Summary: In [my post yesterday](https://fb.workplace.com/groups/657557118131740/permalink/1557733431447433/), I mentioned wanting to test out the effect of using a different isolation dir on the type checking duration and percent of preempted operations. This diff adds that isolation dir and an extra flag in telemetry so that we can filter on this flag.

Reviewed By: MaggieMoss

Differential Revision: D61811369

fbshipit-source-id: 8b863bf54b06a71560cbbf0d1dbc5c3782c82d35