Use non-mangled qualified names for nested funcs in FDE
Summary:
Qualification makes a pretty big mess of the FunctionDefinitionEnvironment data
the way we used to create it, becasue we would take the raw name of each define
directly off the preprocessed AST and use that as a globally unique identifier.

But Pyre doesn't always produce human-readable names in qualfification: global
names are human-readable, but local names are mangled (and not always in the
same way!)

This causes more than one problem:
- Early in his effort on qualification, Jia tried to make a nested define that
  re-binds a parameter use the qualified name of the paramter, which broke Pyre
  incremental (and probably had the potential to break other things) because
  parameter names (after qualification) are not even close to globally unique.
- In addition, the mangled names are not a thing we would want to expose to
  downstream, and we have a client team that wants to be able to use
  `callees_with_location(...)` on nested functions to power their analysis.

In addition to those rather clear-cut problems, there is also the more general
issue that far too much code is trying to destructure the AST of a define into
its name and then use that name as a globally unique key to data stores. This
can never work in a post-qualification code, and quite a lot of the Pysa
codebase is doing it.

----------

This diff piggybacks on Jia's recent addition of `ModuleContext.t` (which
we've tentatively agreed will probably be renamed to `NestingContext.t` at some
point soonish) to the AST which allows us to use *non-mangled* qualified names,
because the human-readable fully qualified name for a define is just the
concatenation of
- the module name
- the concatentated names of all nesting classes and defines
- the (dequalified) name of the define

In the process I create a `FunctionDefinition.qualified_name_of_define`
helper that should produce an unmangled qualified name for a define, and can
be run on both raw and preprocessed ASTs.

-----------

Because this transformation affects *all* lookups in the TypeEnvironment, all
lookups that tried to go straight from AST to a read had to be updated to use
this new helper (because all of them were depending on using `Define.name` for
the lookup, which (a) only works for a qualified AST - so it's great to get
rid of anyway! - and (b) was relying on the mangled name that is no longer
valid).

The diff winds up being slightly large because the fanout (including Pysa,
and including named `Callable` types for local defs which in turn affects
Pysa because Pysa is using the names for its callgraph) is pretty wide.

Since *none* of the updated callsites could have survived the end of qualification
anyway, this diff is a nontrivial step forward on that project as well, since
the new path to get qualified define names is designed to work just as well on
an unqualified or qualified AST.

Some callsites couldn't be updated yet

Reviewed By: grievejia

Differential Revision: D61748415

fbshipit-source-id: 940887b7e3a7ac3aa99ef03e6e3dcb10301815cf