Move variance and constraints data out of `Variable` module
Summary:
In order to model type parameters more cleanly, we need to decouple these
notions from "type variables".

In particular, the variance should be decoupled because by the end of this
stack my intention is to remove it entirely from `Type.Variable.TypeVar.record` -
the variance is never a property of the type var itself, only of the type
parameter on a generic type.

This diff is a pure move / rename, no logical changes or fundamentally new datatypes
here.

**Motivation / what's going to happen in this stack**

This is very important because in our implementation of PEP 695 we will almost
certainly not have computed variance yet at the point where we begin calling
`parse_annotation`, which means it's a significant plubming problem for us to
be assuming "variance is a property of a type var".

Practically, I also need these datatypes extracted early because the
representation of a type parameter (which doesn't exist yet) is going to need
access to these datatypes, and I don't want the `Parameter` module, when I add
it, to live inside of the `Variable` module - they are two related but distinct
concepts and I very much want the datatypes and module names to reflect this.

Actually doing this is going to proceed in some stages:
- trivial moving-around of existing concepts (like this diff) to prep
- making a new datatype to represent Parameters
- using the new datatype in ClassHierarchyEnvironment + downstream
- deleting variance from `Type.Variable.t`

This stack will not actually do anything directly connected to variance inference, but I'm also hoping that once we make the parameter datatype it will be clear what some of our options are (my current thinking is that we should probably have two datatypes - maybe backed by a generic datatype - for "raw" parameters that only know about explicit variance vs parameters after variance inference)

Reviewed By: grievejia

Differential Revision: D61925811

fbshipit-source-id: 97c99fe40bf77db9810a9ef6bb235144948cfa1a