have a lock mechanism to avoid processing the same document in parallel
Summary:
when we do successive saves on a new python file, we end up creating duplicate targets rending the target broken.

with this change, we acquire a lock on the python file that we want to run pyautotargets on and use that as a way to enforce that for each document pyautotargets invocation is sequential.

Reviewed By: zsol

Differential Revision: D61957565

fbshipit-source-id: bd1318a30d79b6a78e625504f4028568541fe64e