Add two unit tests for handling generic methods.
Summary:
In this diff stack, we add support for handling generic functions. We add two unit test for handling generic methods.

One works already and the other one doesn't.

Explanation:

We resolve types correctly in this example:

          T = TypeVar("T")
            class A(Generic[T]):
                def func(self, x: T) -> T:
                    ...

                def func2[U](self, x:U) -> U:
                    ...

The same applies to this example:

            class A:
                def func2[U](self, x:U) -> U:
                    …

However, if we switch to the new syntax for generic classes, then we do not resolve the types for func2

            class A[T]:
                def func(self, x: T) -> T:
                    ...

                def func2[U](self, x:U) -> U:
                    …

In that example, we resolve instantiations a of A[int] for example, as unknown and then a.func2 as
any vs what we would want, which is to resolve both of these types correctly.

So we believe that it is a prerequisite to support generic classes with the new syntax in order to get support for generic methods inside classes that use this new syntax.

Reviewed By: rchen152

Differential Revision: D61890621

fbshipit-source-id: 37b936bb15d8d052a19ecf4935d300340554d89d