Support top level function defintions
Summary:
We adapt our implementation to the fact that we have to seperate scoped type vars from global vars. This is why the implementation looks hacky.

Now we store the mapping in a map instead of defining it using a function, which means in the future we should also think about

1- Mixing scoped type vars with global type vars
2- Nesting and class defintions

We should also find a good solution for handling duplicates. Ideally, I think you can have duplicates and should be able to get the last item you put in the mapping but we can discuss.

Reviewed By: stroxler

Differential Revision: D61870334

fbshipit-source-id: df659ea0e3a231582814b494356f8acb9bf81313