Implement a fallback to global scope if lookup fails in local scope
Summary:
Before our change, this testcase would fail

        from typing import TypeVar, Generic

            T = TypeVar("T")
            class A(Generic[T]):
                def func2[U](self, x:T, y:U) -> T | U:
                    ...

            a = A[int]()
            reveal_type(a.func2)

As it would interpret the type for func2 as `[Unknown, U ]-> Unknown | U`. The reason is that in `parse_annotation`, we do not look at the global scope when the lookup fails. This diff fixes this behavior by adding global lookup as a fallback.

Reviewed By: stroxler

Differential Revision: D61987493

fbshipit-source-id: b0024ec26f5b2f410b676d672319b0296f6be7f9