Split the pyre backend diagram into "overview" and "type checker" diagrams
Summary:
Splits the backend diagram into two chunks so that the images can be made larger for readability. I kept the overview on the overall architecture page and put the new type checker diagram in server.md, which I'm going to take a stab at filling in next.

I've moved some of the boxes around to make the diagrams more compact, but the content should still be the same aside from a couple extra "ast from parser" and "type information and errors" labels I added to show inputs/outputs.

Reviewed By: yangdanny97

Differential Revision: D61986042

fbshipit-source-id: f001417fccf0d5f5f27baf3f4f193364e030d29b