Make `Type.Variable.Declaration.t` primitive in TAE outgoing logic
Summary:
This commit is a pure refactor; we expose the notion of `Type.Variable.Declaration.t`
as the "primitive" form of a type variable in the alias environment, and make
`get_variable` (which parses the declaration, converting bounds to types) an
outgoing computation rather than a core part of the environment itself.

----

**Stack and motivation**

I believe that we should eliminate variance from the datatype `Type.Variable.t`
entirely because at the point when `parse_annotation` is called, we won't yet
have variance inferred for PEP 695 style variables (including legacy variables
using the `infer_variance=True` flag).

There are other BE reasons to try to avoid needing `Type.Variable.t` (e.g. any calls to
`Type.create` that live below `AttributeResolution` add complexity to our system, because
the behavior of `Type.create` in `AttributeResolution` and below is actually different),
but the concern about whether variance "lives on the type itself" is the primary motivatin
here.

This is not conceptually a problem at all because variance isn't a really a property
of type variables in the first place, but it *is* a significant issue with our current
plumbing and data types.

Reviewed By: grievejia

Differential Revision: D61956519

fbshipit-source-id: 5227e48304c939a33a64f4977673358409b114af