Expose `generic_parameters` from classHierarchy
Summary:
I'm not quite ready yet to actually use this yet, but I'm going
to need it to cut variance-related callsites over and now is
a reasonable time to add the interface.

Existing callsites still use `generic_parameters_as_variables`
which is defined as a helper function downstream. In this stack
I only really need to cut over 3 callsites that are
variance-related, which will happen later.

Reviewed By: yangdanny97

Differential Revision: D61994058

fbshipit-source-id: f87f89d66021ba6471c85fb74ca04861d291c528