Create a Type.GenericParameter type
Summary:
This is a step on the road to modeling type parameters explicitly as a concept
that is related to but distinct from type variables - we need some data structure
that isn't a `Type.Variable.t` that we can use to represent generic type constructor
parameters.

I acutally suspect that eventually we'll want at least two variations on this datatype,
one for the parameters before variance inference (which will know about the kinds,
maybe the bounds, and any explicitly-set variance) versus after variance (which will
know about everything), but we don't need to cross that bridge until we actually decide
how to wire variance inference in, so for now I have just one new type.

This diff is just providing the definition, which is the simplest possible definition,
plus some conversions between the existing `Type.Variable.t` and the new type because
we'll need these conversions at least until more plumbing is in place; eventually I
believe that the `of_variable` converter can be removed.

**Context**

I'm working on separating `ClassHierarchyEnvironment`s representation of type parameters
from type variables, because without this PEP 695 support is going to lead to major
confusion around "where variance is stored" - it's currently being stored on type variables
themselves, but we're going to have to run `parse_annotation` at an earlier stage of analysis
than where we can do inference.

This isn't a conceptual problem at all, because variance is really property only of a type
constructor, not of a type variable; I've already confirmed (by deleting the variance and
tracking down compile errors) that we are indeed only actually using variance of type
variables that are "standing in for" generic parameters, where by "standing in for" I mean
we don't have a datatype modeling type parameters right now, and as a result we use
`Type.Variable.t` for that.

So I know that this re-plumbing can work. At a rough guide my plan in this stack is to:
(a) cut over ClassHierarchy to store data in the new `GenericParameter.t` type
(b) cut over `generic_parameters_as_variables` to output a `GenericParameter.t list`
    and cut all the callers over
(c) delete the variance from `Type.Variable.t`

Once this is all done, we will have accomplished two significant things:
- We will have eliminated a potential blocker for 695 (the immediate motivation)
- We also will have modeled the concept of a `GenericParameter` explicitly in our
  codebase, as a distinct concept from a `Variable` (as well as a distinct concept
  from a type `Argument` - the three are related but not the same!) which is probably
  a significant BE win in the long run.
  - Note that we will *not* have modeled the concept of a "for all parameter" as Pyre
    currently deals with it in generic functions; at the end of the stack those are
    probably still going to be implicitly modeled via the various `Type.Variable.t`
    values showing up in a signature.
  - Whether these two concepts are fundamentally different is open to debate (I think
    at the moment you could model the spec in more than one way) but they are
    definitely different in the current codebase, they appear in 100% disjoint
    places.
  - The existing modeling of generic functions is probably a significant issue
    from a better engineering point of view, but I don't believe it is a blocker for
    695 support because the variance inference problem isn't relevant. That's
    why I'm not planning to address it at this time.

Reviewed By: grievejia

Differential Revision: D61994060

fbshipit-source-id: b97824f93c1e8452b8502a56f84330923d38a7d8